move resolve_callables into utils.py
